refactor list and watch logic out
